i18n the group view
