+++ webapp/src/components/group/GroupTable.tsx
@@ -1,6 +1,7 @@
 import  {  useEffect } from 'react';
 import { Table, TableBody, TableCell, TableContainer, TableHead, TableRow, Container, Grid, Button } from "@mui/material"
 import axios from 'axios';
+import { useTranslation } from 'react-i18next';
 
 type TableProps = {
     groupUUID: string,
@@ -24,6 +25,7 @@ let numberMembers = 0;
 const apiEndpoint = process.env.REACT_APP_API_ENDPOINT || 'http://localhost:8000';
 
 export const GroupTable = (props: TableProps) => {
+    const { t } = useTranslation();
     const aFunction = async ()=>{
         await axios.get(`${apiEndpoint}/getGroup/`+props.groupUUID).then(res => {
             console.log(res.data);
@@ -31,9 +33,9 @@ export const GroupTable = (props: TableProps) => {
             numberMembers=0;
             total = 0;
             for(let member of res.data.members){
-                let memberRole = "Member";
+                let memberRole = t('group_table_member');
                 if(member.uuid == res.data.admin.uuid){
-                    memberRole = "Leader";
+                    memberRole = t('group_table_leader');
                 }
                 members.push({
                     username : member.username,
@@ -91,9 +93,9 @@ export const GroupTable = (props: TableProps) => {
                 <Table>
                     <TableHead>
                         <TableRow>
-                            <TableCell>Username</TableCell>
-                            <TableCell>Role</TableCell>
-                            <TableCell>Score</TableCell>
+                            <TableCell>{t('group_table_username')}</TableCell>
+                            <TableCell>{t('group_table_role')}</TableCell>
+                            <TableCell>{t('group_table_score')}</TableCell>
                         </TableRow>
                     </TableHead>
                     <TableBody>
