+++ webapp/src/components/group/NoGroup.tsx
@@ -1,7 +1,8 @@
 import './Group.scss';
 import { Button, Container, Snackbar, TextField, Grid, Stack, RadioGroup, FormControlLabel, Radio } from "@mui/material";
-import  { ChangeEvent, useEffect, useState } from 'react';
+import  { useEffect, useState } from 'react';
 import axios from 'axios';
+import { CreationModal } from './GroupCreationModal';
 
 const apiEndpoint = process.env.REACT_APP_API_ENDPOINT || 'http://localhost:8000';
 
@@ -24,13 +25,9 @@ const NoGroup = (props: ActionProps) =>
     const [error, setError] = useState('');
     const [createModal, setCreateModal] = useState(false);
     const [joinModal, setJoinModal] = useState(false);
-    const [groupName, setGroupName] = useState('');
-    const [isPublic, setPublic] = useState(true);
-    const [maxMembers, setMaxMembers] = useState(2);
-    const [description, setDescription] = useState('');
+   
     const creatorUUID =  JSON.stringify(localStorage.getItem("userUUID")).replace("\"", "").replace("\"", "");
     
-
     const toggleCreateModal = () => {
         setCreateModal(!createModal);
     };
@@ -41,19 +38,7 @@ const NoGroup = (props: ActionProps) =>
             findGroups();
         }
     };
-
-    const createGroup = async () =>{
-        try{
-            console.log("Public?");
-            console.log(isPublic);
-            await axios.post(`${apiEndpoint}/createGroup`, { groupName, creatorUUID, description, isPublic }).then( res => {
-                props.nowHasGroup();
-            });
-        }catch (error:any) {
-        setError(error.response.data.error);
-        }
-    }
-
+    
     useEffect(() => {
         findGroups();
     })
@@ -96,18 +81,7 @@ const NoGroup = (props: ActionProps) =>
         }
     }
 
-    const handleChange = ({ currentTarget: { value } }: ChangeEvent<HTMLInputElement>) => {
-        setMaxMembers(curr => {
-            if (!Boolean(value)) { return 0; }
-            const numeric = parseInt(value, 10);
-
-            if (value.length > 100) {
-                return curr;
-            }
-
-            return (value.length <= 100 ? numeric : curr);
-        });
-    };
+    
     
     
     return (
@@ -121,55 +95,9 @@ const NoGroup = (props: ActionProps) =>
             {error && (
             <Snackbar open={!!error} autoHideDuration={6000} onClose={() => setError('')} message={`Error: ${error}`} />
             )}
-            {createModal && (
-                <div className="modal">
-                    <div className="modal-content">
-                        <h2>Create group</h2>
-                        <Grid >
-                            <Grid container padding={2} >
-                                <Grid item xs={5} ><p>Group name:</p></Grid>
-                                <Grid item xs={5} ><TextField
-                                margin="normal"
-                                label="Group name"
-                                value={groupName}
-                                onChange={(e) => setGroupName(e.target.value)}
-                                /></Grid>
-                            </Grid>
-                            <Grid container padding={2} >
-                                <Grid item xs={5} ><p>Group name:</p></Grid>
-                                <Grid item xs={5} ><RadioGroup
-                                aria-labelledby="demo-radio-buttons-group-label"
-                                defaultValue="yes"
-                                name="radio-buttons-group"
-                                onChange={(e) => setPublic(e.target.value === "yes")}
-                                >
-                                    <FormControlLabel value="yes" control={<Radio />} label="Yes" />
-                                    <FormControlLabel value="no" control={<Radio />} label="No" />
-                                </RadioGroup></Grid>
-                            </Grid>
-                            <Grid container padding={2} >
-                                <Grid item xs={5} ><p>Max members:</p></Grid>
-                                <Grid item xs={5} ><input type="number" step={1} value={maxMembers} onChange={handleChange} max={200} min={2} /></Grid>
-                            </Grid>
-                            <Grid container padding={2} >
-                                <Grid item xs={5} ><p>Description:</p></Grid>
-                                <Grid item xs={5} ><TextField
-                                margin="normal"
-                                multiline
-                                label="Description"
-                                value={description}
-                                onChange={(e) => setDescription(e.target.value)}
-                                
-                                /></Grid>
-                            </Grid>
-                            <Grid container padding={2} >
-                                <Grid item xs={6} ><Button onClick={toggleCreateModal}>Close</Button></Grid>
-                                <Grid item xs={6} ><Button onClick={createGroup}>Create group</Button></Grid>
-                            </Grid>
-                        </Grid>
-                    </div>
-                </div>
-            )}
+            {createModal &&
+                (<CreationModal nowHasGroup={props.nowHasGroup} setError={setError} closeModal={toggleCreateModal}/>)
+            }
             {joinModal && (groupsCharged && (
                 <div className="modal">
                     <div className="modal-content">
